fix lint issues
